+++ SRC/Private/Comparers/ServiceReferenceComparer.cs
@@ -1,17 +1,13 @@
 ﻿/********************************************************************************
-* ServiceIdExtensions.cs                                                        *
+* ServiceReferenceComparer.cs                                                   *
 *                                                                               *
 * Author: Denes Solti                                                           *
 ********************************************************************************/
+
 namespace Solti.Utils.DI.Internals
 {
-    internal static class ServiceIdExtensions
+    internal sealed class ServiceReferenceComparer : ComparerBase<ServiceReferenceComparer, AbstractServiceReference>
     {
-        public static string FriendlyName(this IServiceID src) 
-        {
-            string result = src.Interface.ToString();
-            if (src.Name != null) result += $":{src.Name}";
-            return result;
-        }
+        public override int GetHashCode(AbstractServiceReference obj) => obj.RelatedServiceEntry.GetHashCode();
     }
 }