+++ SRC/Private/ServiceID.cs
@@ -1,17 +1,16 @@
 ﻿/********************************************************************************
-* ServiceIdExtensions.cs                                                        *
+* ServiceID.cs                                                                  *
 *                                                                               *
 * Author: Denes Solti                                                           *
 ********************************************************************************/
+using System;
+
 namespace Solti.Utils.DI.Internals
 {
-    internal static class ServiceIdExtensions
+    internal class ServiceID : IServiceID
     {
-        public static string FriendlyName(this IServiceID src) 
-        {
-            string result = src.Interface.ToString();
-            if (src.Name != null) result += $":{src.Name}";
-            return result;
-        }
+        public Type Interface { get; set; }
+
+        public string Name { get; set; }
     }
 }