+++ SRC/Public/Exceptions/CircularReferenceException.cs
@@ -23,13 +23,13 @@ namespace Solti.Utils.DI
         /// Creates a new <see cref="CircularReferenceException"/> instance.
         /// </summary>
         /// <param name="path">The current path on which the circular reference was found.</param>
-        public CircularReferenceException(IEnumerable<ServiceReference> path) : base
+        public CircularReferenceException(IEnumerable<AbstractServiceReference> path) : base
         (
             string.Format
             (
                 Resources.Culture,
                 Resources.CIRCULAR_REFERENCE,
-                string.Join(" -> ", path.Select(part => (part.Interface, part.Name).FriendlyName()))
+                string.Join(" -> ", path.Select(part => part.RelatedServiceEntry.FriendlyName()))
             )
         ) 
         {