+++ TEST/Injector/Get.cs
@@ -9,6 +9,7 @@ using NUnit.Framework;
 
 namespace Solti.Utils.DI.Injector.Tests
 {
+    using Internals;
     using Properties;
     
     public partial class InjectorTestsBase<TContainer>
@@ -85,6 +86,8 @@ namespace Solti.Utils.DI.Injector.Tests
         [TestCase(Lifetime.Singleton)]
         public void Injector_Get_ShouldResolveGenericDependencies(Lifetime lifetime)
         {
+            Config.Value.Injector.StrictDI = false;
+
             Container
                 .Service<IInterface_1, Implementation_1_No_Dep>() // direkt nincs lifetime
                 .Service(typeof(IInterface_3<>), typeof(Implementation_3_IInterface_1_Dependant<>)) // direkt nincs lifetime
@@ -185,6 +188,17 @@ namespace Solti.Utils.DI.Injector.Tests
             using (IInjector injector = Container.CreateInjector()) 
             {
                 Assert.Throws<Exception>(() => injector.Get<IInterface_1>(), string.Format(Resources.INVALID_INSTANCE, typeof(IInterface_1)));
+            }
+        }
+
+        [Test]
+        public void Injector_Get_ShouldThrowOnAbstractService() 
+        {
+            using (IInjector injector = Container.CreateInjector()) 
+            {
+                injector.UnderlyingContainer.Abstract<IInterface_1>();
+
+                Assert.Throws<InvalidOperationException>(() => injector.Get<IInterface_1>(), Resources.CANT_INSTANTIATE_ABSTRACTS);
             }
         }
     }