+++ TEST/Injector/Lifetime.cs
@@ -28,7 +28,7 @@ namespace Solti.Utils.DI.Injector.Tests
         [Test]
         public void Lifetime_TransientService_ShouldNotBeInstantiatedIfTheInjectorWasRecycled() 
         {
-            Config.Value.MaxSpawnedServices = 1;
+            Config.Value.Injector.MaxSpawnedTransientServices = 1;
 
             Container.Service<IInterface_1, Implementation_1_No_Dep>(Lifetime.Transient);
 
@@ -117,6 +117,8 @@ namespace Solti.Utils.DI.Injector.Tests
         [Test]
         public void Lifetime_SingletonService_MayHaveScopedDependency()
         {
+            Config.Value.Injector.StrictDI = false;
+
             Disposable instance;
 
             using (IServiceContainer child = Container.CreateChild())
@@ -138,6 +140,27 @@ namespace Solti.Utils.DI.Injector.Tests
             }
 
             Assert.That(instance.Disposed);
+        }
+
+        [Test]
+        public void Lifetime_SingletonService_ShouldHaveItsOwnInjector() 
+        {
+            Container
+                .Service<IInterface_7<IInjector>, Implementation_7_TInterface_Dependant<IInjector>>(Lifetime.Singleton)
+                .Service<IInterface_7<IInjector>, Implementation_7_TInterface_Dependant<IInjector>>("named", Lifetime.Singleton);
+
+            using (IInjector injector = Container.CreateInjector()) 
+            {
+                IInterface_7<IInjector> svc = injector.Get<IInterface_7<IInjector>>("named");
+
+                Assert.That(svc.Interface, Is.Not.SameAs(injector));
+                Assert.That(svc.Interface.UnderlyingContainer.Parent, Is.SameAs(Container));
+
+                Assert.That(svc.Interface.Get<IInterface_7<IInjector>>().Interface, Is.Not.SameAs(injector));
+                Assert.That(svc.Interface.Get<IInterface_7<IInjector>>().Interface.UnderlyingContainer.Parent, Is.SameAs(Container));
+
+                Assert.That(svc.Interface.Get<IInterface_7<IInjector>>().Interface, Is.Not.SameAs(svc.Interface));
+            }
         }
 
         [TestCase(Lifetime.Transient)]
@@ -172,6 +195,8 @@ namespace Solti.Utils.DI.Injector.Tests
         [Test]
         public void Lifetime_SingletonService_ShouldResolveDependencyFromTheDeclaringContainer_DecorationTest()
         {
+            Config.Value.Injector.StrictDI = false;
+
             Container
                 .Service<IInterface_1, Implementation_1_No_Dep>(Lifetime.Transient)
                 .Service<IInterface_2, Implementation_2_IInterface_1_Dependant>(Lifetime.Singleton);
@@ -199,6 +224,121 @@ namespace Solti.Utils.DI.Injector.Tests
 
                 Assert.AreSame(injector1.UnderlyingContainer.Get<IDisposable>(), injector2.UnderlyingContainer.Get<IDisposable>());
                 Assert.AreSame(injector1.Get<IDisposable>(), injector2.Get<IDisposable>());
+            }
+        }
+
+        [Test]
+        public void Lifetime_PermissiveDI_LegalCases(
+            [Values(true, false)] bool useChildContainer,
+            [Values(Lifetime.Transient, Lifetime.Scoped, Lifetime.Singleton)] Lifetime dependant,
+            [Values(Lifetime.Transient, Lifetime.Scoped, Lifetime.Singleton, null)] Lifetime? dependency)
+        {
+            Config.Value.Injector.StrictDI = false;
+
+            if (dependency != null)
+                Container.Service<IInterface_1, Implementation_1_No_Dep>(dependency.Value);
+            else
+                Container.Instance<IInterface_1>(new Implementation_1_No_Dep());
+
+            IServiceContainer sourceContainer = useChildContainer ? Container.CreateChild() : Container;
+
+            sourceContainer.Service<IInterface_2, Implementation_2_IInterface_1_Dependant>(dependant);
+
+            //
+            // Ket kulonallo injectort hozzunk letre.
+            //
+
+            for (int i = 0; i < 2; i++)
+            {
+                using (IInjector injector = sourceContainer.CreateInjector())
+                {
+                    Assert.DoesNotThrow(() => injector.Get<IInterface_2>());
+                }
+            }
+        }
+
+        [Test]
+        public void Lifetime_StrictDI_LegalCases1(
+            [Values(true, false)] bool useChildContainer, 
+            [Values(Lifetime.Transient, Lifetime.Scoped)] Lifetime dependant, 
+            [Values(Lifetime.Transient, Lifetime.Scoped, Lifetime.Singleton, null)] Lifetime? dependency) 
+        {
+            Config.Value.Injector.StrictDI = true;
+
+            if (dependency != null)
+                Container.Service<IInterface_1, Implementation_1_No_Dep>(dependency.Value);
+            else
+                Container.Instance<IInterface_1>(new Implementation_1_No_Dep());
+
+            Container.Service<IInterface_2, Implementation_2_IInterface_1_Dependant>(dependant);
+
+            IServiceContainer sourceContainer = useChildContainer ? Container.CreateChild() : Container;
+
+            //
+            // Ket kulonallo injectort hozzunk letre.
+            //
+
+            for (int i = 0; i < 2; i++)
+            {
+                using (IInjector injector = sourceContainer.CreateInjector())
+                {
+                    Assert.DoesNotThrow(() => injector.Get<IInterface_2>());
+                }
+            }
+        }
+
+        [Test]
+        public void Lifetime_StrictDI_LegalCases2(
+            [Values(true, false)] bool useChildContainer,
+            [Values(Lifetime.Singleton, null)] Lifetime? dependency)
+        {
+            Config.Value.Injector.StrictDI = true;
+
+            if (dependency != null)
+                Container.Service<IInterface_1, Implementation_1_No_Dep>(dependency.Value);
+            else
+                Container.Instance<IInterface_1>(new Implementation_1_No_Dep());
+
+            Container.Service<IInterface_2, Implementation_2_IInterface_1_Dependant>(Lifetime.Singleton);
+
+            IServiceContainer sourceContainer = useChildContainer ? Container.CreateChild() : Container;
+
+            //
+            // Ket kulonallo injectort hozzunk letre.
+            //
+
+            for (int i = 0; i < 2; i++)
+            {
+                using (IInjector injector = sourceContainer.CreateInjector())
+                {
+                    Assert.DoesNotThrow(() => injector.Get<IInterface_2>());
+                }
+            }
+        }
+
+        [Test]
+        public void Lifetime_StrictDI_IllegalCases(
+            [Values(true, false)] bool useChildContainer,
+            [Values(Lifetime.Transient, Lifetime.Scoped)] Lifetime dependency) 
+        {
+            Config.Value.Injector.StrictDI = true;
+
+            Container
+                .Service<IInterface_1, Implementation_1_No_Dep>(dependency)
+                .Service<IInterface_2, Implementation_2_IInterface_1_Dependant>(Lifetime.Singleton);
+
+            IServiceContainer sourceContainer = useChildContainer ? Container.CreateChild() : Container;
+
+            //
+            // Ket kulonallo injectort hozzunk letre.
+            //
+
+            for (int i = 0; i < 2; i++)
+            {
+                using (IInjector injector = sourceContainer.CreateInjector())
+                {
+                    Assert.Throws<RequestNotAllowedException>(() => injector.Get<IInterface_2>());
+                }
             }
         }
     }